legg spinner på vurderingknapper & rename fil
