+++ src/pages/saksbehandling/søknadsbehandling/flyktning/Flyktning.tsx
@@ -29,7 +29,7 @@ import { erUnderkjent, erVilkårsvurderingerVurdertAvslag } from '~src/utils/beh
 import sharedI18n from '../sharedI18n-nb';
 import * as sharedStyles from '../sharedStyles.module.less';
 import { VilkårsvurderingBaseProps } from '../types';
-import { Vurderingknapper } from '../Vurdering';
+import { Vurderingknapper } from '../Vurderingknapper';
 
 import messages from './flyktning-nb';
