+++ src/pages/saksbehandling/søknadsbehandling/opphold-i-utlandet/OppholdIUtlandet.tsx
@@ -27,7 +27,7 @@ import { sluttenAvMåneden, toIsoDateOnlyString } from '~src/utils/date/dateUtil
 import sharedI18n from '../sharedI18n-nb';
 import * as sharedStyles from '../sharedStyles.module.less';
 import { VilkårsvurderingBaseProps } from '../types';
-import { Vurderingknapper } from '../Vurdering';
+import { Vurderingknapper } from '../Vurderingknapper';
 
 import messages from './oppholdIUtlandet-nb';
