+++ src/pages/saksbehandling/søknadsbehandling/personlig-oppmøte/PersonligOppmøte.tsx
@@ -33,7 +33,7 @@ import { Vilkårsinformasjon, mapToVilkårsinformasjon } from '~src/utils/søkna
 
 import sharedI18n from '../sharedI18n-nb';
 import { VilkårsvurderingBaseProps } from '../types';
-import { Vurderingknapper } from '../Vurdering';
+import { Vurderingknapper } from '../Vurderingknapper';
 
 import messages from './personligOppmøte-nb';
 import * as styles from './personligOppmøte.module.less';
